api/requirements.txt

Flask==1.1.1
Flask-Admin==1.5.3
Flask-AppBuilder==1.13.1
Flask-Babel==0.12.2
Flask-Caching==1.3.3
Flask-JWT-Extended==3.23.0
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.1
flask-swagger==0.2.13
Flask-WTF>=0.14.3
ipython==7.12.0
ipython-genutils==0.2.0
Jinja2==2.11.1
json-merge-patch==0.2
json5==0.9.1
jsonschema==3.2.0
python-crfsuite==0.9.6
python-daemon==2.1.2
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
python-utils==2.3.0
python3-openid==3.1.0
regex==2019.8.19
requests==2.22.0
requests-oauthlib==1.3.0
BeautifulSoup4
SQLAlchemy==1.3.13
sqlparse==0.3.0
urllib3==1.25.8
PyMySQL
Flask-Migrate
threaded
schedule
APScheduler
email-validator
flask-bootstrap4
Flask-Cors